Renamed base package. Fixed NullExceptionPointer bug in DashboardActivityViewModel
