+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.example.cavoid"
+        applicationId "com.operationcodify.cavoid"
         minSdkVersion 27
         targetSdkVersion 30
         versionCode 1
