+++ app/src/main/java/com/operationcodify/cavoid/database/PastLocation.java
@@ -1,0 +1,31 @@
+package com.operationcodify.cavoid.database;
+
+import androidx.annotation.NonNull;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+
+import org.joda.time.Instant;
+import org.joda.time.LocalDate ;
+
+@Entity(tableName = "past_location",
+        primaryKeys = {"date", "fips"}
+
+)
+public class PastLocation {
+
+    @NonNull
+    @ColumnInfo(name = "date")
+    public LocalDate date;
+
+    @NonNull
+    @ColumnInfo(name = "fips")
+    public String fips;
+
+    @ColumnInfo(name="county_name")
+    public String countyName;
+
+    @ColumnInfo(name = "was_notified")
+    public Boolean wasNotified = false;
+
+    public long timestamp = Instant.now().getMillis();
+}
