+++ app/src/main/java/com/operationcodify/cavoid/utilities/PastLocationViewHolder.java
@@ -1,0 +1,17 @@
+package com.operationcodify.cavoid.utilities;
+
+import android.view.View;
+import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView.ViewHolder;
+
+import com.operationcodify.cavoid.R;
+
+public class PastLocationViewHolder extends ViewHolder {
+    TextView pastCasesTextView,pastDeathsTextView;
+    public PastLocationViewHolder(View v){
+        super(v);
+        this.pastCasesTextView = v.findViewById(R.id.pastCasesTextView);
+        this.pastDeathsTextView = v.findViewById(R.id.pastDeathsTextView);
+    }
+}
